Implement type erasure system for walk()
The design of the Filesystem Adaptor interface required a type
erasure system to allow the IFilesystemAdaptor::walk method to be
decoupled from the container it receives (generic) and polymorphic.
The method itself is still unimplemented, but the test and CMake
files have been updated to conform to the new interface.