Stuf out TagLibAdaptor class
The TagLibAdaptor implements the IMusicTagEditorAdaptor interface.
MusicFiles will receive a reference to an object derived of
IMusicTagEditorAdaptor, and will interact with this object to set
and read tags and other metadata in the music file. TagLib will be
able to support all the Music file types that are currently in the
design doc.