Implement OrganizationalPolicy Framework
OrganizationalPolicies outline requirements that media files in an
organized library must meet. Policies may depend on other policies,
But Organizational Schemas (upcoming commit) must match exactly one
file type to exactly one policy. The IOrganizationalPolicy interface
has two methods: `verify', which is used to determine that all the
prerequisites necessary to adminiter an Organizational Policy have
been met (so we can know early whether or not a file qualifies for
organization under the current Schema). Finally, there is
`administer', which is used to determine the operations necessary
to ensure that the file conforms to the policy.