Implement OrganizationalSchema Framework
Organizational Schemas are factory-like classes that match file
types to OrganizationalPolicies. The Abstract base class
OrganizationalSchema is also responsible for extracting the
IFileOperations that need to be applied to the file and applying
them. Eventually, after a robust exception hierarchy is in place,
the OrganizationalSchema will also be aware of how to revert
operations in the case of an error.