Patch default schemas and policies into MediaSorter
In this commit, the default or "basic" schemas and policies utilizing
the currently implemented operations were patched into the factory
classes and Media Sorter. This patch also includes some cursory bug
fixes in WarnOperation and the ArgumentParser class. Currently,
BasicMusicOrganizationalSchema is experiencing a bug due to a logical
error in FSAdaptor::Path. This will be fixed in the FSAdaptor library
upstream, and then brought into master in a future commit.