Enforce "track total" and update padding
This patch enforces the need for a "track total" field which is
taken from the track number field. The format for the tracknumber
field is now "<num>/<total>." Update the zero-padding algorithm to
use this information for padding. This came as a result of a bug
that was discovered when trying to sort albums that had more than
99 tracks. Unfortunately, every music file must follow this
convention now.