Refactor adaptor to use command/key pattern
Previously, children of IMusicTagEditorAdaptor implemented a large
interface with many getters and setters, one for each attribute.
This became unmaintainable and not flexible for polymorphism, so
this class was refactored to use command pattern, with only one
'get' and one 'set' method which are used to get and set all tags,
respectively. The tag is specified with the first argument to these
functions, which is a member of an enum declared in
IMusicTagEditorAdaptor. The consumers of these classes were also
refactored.