Add Disc number to music filenames
If the disc number is set, add the disc number to the front of the
filename in BasicMusicFileNamingPolicy, zero padded to have a
consistent string width for all files. If the disc number is set,
but the disc total is not, issue a warning to the user. This works
for all supported music file formats.