Exception Handling
If a child of TagLibAdaptor does not implement a getter/setter for
a particular tag, the program will exit with a message. Also update
handling of exceptions in general--always catch const, and a main
try/catch was created in main.cpp to ensure that a message is
printed if an exception makes it all the way to main.