Fix warnings about std::iterator deprecation
std::iterator was deprecated in C++17, because it provides little
clarity for custom iterator definitions. Besides, it was only a
convenience for the more verbose using declarations that normally
identify an iterator. Move PathConstIter to be a member class of
Path, and create the `using` decl tags to specify this as an
iterator instead. For more info, see:
https://www.internalpointers.com/post/writing-custom-iterators-modern-cpp